+++ laoschoolws/src/main/java/com/itpro/restws/dao/MExamDaoImpl.java
@@ -21,7 +21,7 @@ public class MExamDaoImpl extends AbstractDao<Integer, MExam> implements MExamDa
 	@Override
 	public int countBySchool(Integer school_id) {
 		// Get row count
-		int count = ((Integer)getSession().createQuery("select count(*) from " + ModelName+  " WHERE school_id = '" + school_id + "'").uniqueResult()).intValue();
+		int count = ((Long)getSession().createQuery("select count(*) from " + ModelName+  " WHERE school_id = '" + school_id + "'").uniqueResult()).intValue();
 		return count;
 	}
