+++ laoschoolws/src/main/java/com/itpro/restws/dao/MSubjectDaoImpl.java
@@ -24,7 +24,7 @@ public class MSubjectDaoImpl extends AbstractDao<Integer, MSubject> implements M
 	@Override
 	public int countBySchool(Integer school_id) {
 		// Get row count
-		int count = ((Integer)getSession().createQuery("select count(*) from " + ModelName+  " WHERE school_id = '" + school_id + "'").uniqueResult()).intValue();
+		int count = ((Long)getSession().createQuery("select count(*) from " + ModelName+  " WHERE school_id = '" + school_id + "'").uniqueResult()).intValue();
 		return count;
 	}
