+++ laoschoolws/src/main/java/com/itpro/restws/dao/SysDegreeDaoImpl.java
@@ -28,7 +28,7 @@ public class SysDegreeDaoImpl extends AbstractDao<Integer, SysDegree> implements
 	public int countAll() {
 		
 		// Get row count
-		int count = ((Integer)getSession().createQuery("select count(*) from " + SysTblName.TBLNAME_SYS_DEGREE.getModelName()+  " WHERE actflg = 'A'").uniqueResult()).intValue();
+		int count = ((Long)getSession().createQuery("select count(*) from " + SysTblName.TBLNAME_SYS_DEGREE.getModelName()+  " WHERE actflg = 'A'").uniqueResult()).intValue();
 		return count;
 
 	}
